# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Event-Offer Management API","description":"A FastAPI application for managing Events, Offers, and their associations","version":"1.0.0"},"paths":{"/api/events/":{"post":{"tags":["events-offers"],"summary":"Create New Event","operationId":"create_new_event_api_events__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["events-offers"],"summary":"Read Events","operationId":"read_events_api_events__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Event"},"title":"Response Read Events Api Events  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/events/{event_id}":{"get":{"tags":["events-offers"],"summary":"Read Event","operationId":"read_event_api_events__event_id__get","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["events-offers"],"summary":"Update Existing Event","operationId":"update_existing_event_api_events__event_id__put","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["events-offers"],"summary":"Delete Existing Event","operationId":"delete_existing_event_api_events__event_id__delete","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/offers/":{"post":{"tags":["events-offers"],"summary":"Create New Offer","operationId":"create_new_offer_api_offers__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfferCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Offer"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["events-offers"],"summary":"Read Offers","operationId":"read_offers_api_offers__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Offer"},"title":"Response Read Offers Api Offers  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/offers/{offer_id}":{"get":{"tags":["events-offers"],"summary":"Read Offer","operationId":"read_offer_api_offers__offer_id__get","parameters":[{"name":"offer_id","in":"path","required":true,"schema":{"type":"string","title":"Offer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Offer"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["events-offers"],"summary":"Update Existing Offer","operationId":"update_existing_offer_api_offers__offer_id__put","parameters":[{"name":"offer_id","in":"path","required":true,"schema":{"type":"string","title":"Offer Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfferUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Offer"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["events-offers"],"summary":"Delete Existing Offer","operationId":"delete_existing_offer_api_offers__offer_id__delete","parameters":[{"name":"offer_id","in":"path","required":true,"schema":{"type":"string","title":"Offer Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/event-offers/":{"post":{"tags":["events-offers"],"summary":"Create Event Offer Link","operationId":"create_event_offer_link_api_event_offers__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventOfferCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventOfferCreate"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/event-offers/{event_id}/{offer_id}":{"put":{"tags":["events-offers"],"summary":"Update Event Offer Link","operationId":"update_event_offer_link_api_event_offers__event_id___offer_id__put","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}},{"name":"offer_id","in":"path","required":true,"schema":{"type":"string","title":"Offer Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventOfferUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventOfferUpdate"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["events-offers"],"summary":"Delete Event Offer Link","operationId":"delete_event_offer_link_api_event_offers__event_id___offer_id__delete","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}},{"name":"offer_id","in":"path","required":true,"schema":{"type":"string","title":"Offer Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Event":{"properties":{"code":{"type":"string","title":"Code"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"lifetime_hours":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Lifetime Hours","default":0},"disable_all_campaigns":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Disable All Campaigns","default":false},"queue_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Queue Name","default":"default_queue"},"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["code","name","id","created_at","updated_at"],"title":"Event"},"EventCreate":{"properties":{"code":{"type":"string","title":"Code"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"lifetime_hours":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Lifetime Hours","default":0},"disable_all_campaigns":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Disable All Campaigns","default":false},"queue_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Queue Name","default":"default_queue"},"id":{"type":"string","title":"Id"}},"type":"object","required":["code","name","id"],"title":"EventCreate"},"EventOfferCreate":{"properties":{"event_id":{"type":"string","title":"Event Id"},"offer_id":{"type":"string","title":"Offer Id"},"delay_minutes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Delay Minutes","default":0},"action_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Action Type","default":"enable"}},"type":"object","required":["event_id","offer_id"],"title":"EventOfferCreate"},"EventOfferUpdate":{"properties":{"delay_minutes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Delay Minutes"},"action_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Action Type"}},"type":"object","title":"EventOfferUpdate"},"EventUpdate":{"properties":{"code":{"type":"string","title":"Code"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"lifetime_hours":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Lifetime Hours","default":0},"disable_all_campaigns":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Disable All Campaigns","default":false},"queue_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Queue Name","default":"default_queue"}},"type":"object","required":["code","name"],"title":"EventUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Offer":{"properties":{"code":{"type":"string","title":"Code"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"priority":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Priority","default":1},"target_system":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target System","default":"default_system"},"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["code","name","id","created_at","updated_at"],"title":"Offer"},"OfferCreate":{"properties":{"code":{"type":"string","title":"Code"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"priority":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Priority","default":1},"target_system":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target System","default":"default_system"},"id":{"type":"string","title":"Id"}},"type":"object","required":["code","name","id"],"title":"OfferCreate"},"OfferUpdate":{"properties":{"code":{"type":"string","title":"Code"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"priority":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Priority","default":1},"target_system":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target System","default":"default_system"}},"type":"object","required":["code","name"],"title":"OfferUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}